backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
scikit-image==0.22.0
torch==2.2.0
torchvision==0.17.0
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
